M3-116 Refactor : trailing comma 추가
